update affiliation connection loaders for ordering
