+++ api-js/src/affiliation/loaders/index.js
@@ -1,3 +1,3 @@
-export * from './load-affiliations-by-key'
-export * from './load-affiliations-by-org-id'
-export * from './load-affiliations-by-user-id'
+export * from './load-affiliation-by-key'
+export * from './load-affiliation-connections-by-org-id'
+export * from './load-affiliation-connections-by-user-id'
