+++ api-js/src/affiliation/loaders/load-affiliation-connections-by-org-id.js
@@ -2,15 +2,13 @@ import { aql } from 'arangojs'
 import { fromGlobalId, toGlobalId } from 'graphql-relay'
 import { t } from '@lingui/macro'
 
-export const affiliationLoaderByOrgId = (
+export const affiliationConnectionLoaderByOrgId = (
   query,
-  language,
   userKey,
   cleanseInput,
   i18n,
 ) => async ({ orgId, after, before, first, last, orderBy }) => {
   let afterTemplate = aql``
-
   if (typeof after !== 'undefined') {
     const { id: afterId } = fromGlobalId(cleanseInput(after))
     if (typeof orderBy === 'undefined') {
@@ -26,62 +24,14 @@ export const affiliationLoaderByOrgId = (
       let affiliationField, documentField
       /* istanbul ignore else */
       if (orderBy.field === 'user-username') {
-        affiliationField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).username`
-        documentField = aql`DOCUMENT(users, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._to).key).username`
-      } else if (orderBy.field === 'org-acronym') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).acronym`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).acronym`
-      } else if (orderBy.field === 'org-name') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).name`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).name`
-      } else if (orderBy.field === 'org-slug') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).slug`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).slug`
-      } else if (orderBy.field === 'org-zone') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).zone`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).zone`
-      } else if (orderBy.field === 'org-sector') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).sector`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).sector`
-      } else if (orderBy.field === 'org-country') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).country`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).country`
-      } else if (orderBy.field === 'org-province') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).province`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).province`
-      } else if (orderBy.field === 'org-city') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).city`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).orgDetails).city`
-      } else if (orderBy.field === 'org-verified') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).verified`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).verified`
-      } else if (orderBy.field === 'org-summary-mail-pass') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.pass`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.mail.pass`
-      } else if (orderBy.field === 'org-summary-mail-fail') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.fail`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.mail.fail`
-      } else if (orderBy.field === 'org-summary-mail-total') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.total`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.mail.total`
-      } else if (orderBy.field === 'org-summary-web-pass') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.pass`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.web.pass`
-      } else if (orderBy.field === 'org-summary-web-fail') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.fail`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.web.fail`
-      } else if (orderBy.field === 'org-summary-web-total') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.total`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key).summaries.web.total`
-      } else if (orderBy.field === 'org-domain-count') {
-        affiliationField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key)._id claims RETURN e._to)`
-        documentField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._from).key)._id claims RETURN e._to)`
+        affiliationField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).userName`
+        documentField = aql`DOCUMENT(users, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${afterId})._to).key).userName`
       }
 
       afterTemplate = aql`
         FILTER ${affiliationField} ${afterTemplateDirection} ${documentField}
         OR (${affiliationField} == ${documentField}
-        AND FILTER TO_NUMBER(affiliation._key) > TO_NUMBER(${afterId}))
+        AND TO_NUMBER(affiliation._key) > TO_NUMBER(${afterId}))
       `
     }
   }
@@ -102,62 +52,14 @@ export const affiliationLoaderByOrgId = (
       let affiliationField, documentField
       /* istanbul ignore else */
       if (orderBy.field === 'user-username') {
-        affiliationField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).username`
-        documentField = aql`DOCUMENT(users, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._to).key).username`
-      } else if (orderBy.field === 'org-acronym') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).acronym`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).acronym`
-      } else if (orderBy.field === 'org-name') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).name`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).name`
-      } else if (orderBy.field === 'org-slug') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).slug`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).slug`
-      } else if (orderBy.field === 'org-zone') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).zone`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).zone`
-      } else if (orderBy.field === 'org-sector') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).sector`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).sector`
-      } else if (orderBy.field === 'org-country') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).country`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).country`
-      } else if (orderBy.field === 'org-province') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).province`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).province`
-      } else if (orderBy.field === 'org-city') {
-        affiliationField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).city`
-        documentField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).orgDetails).city`
-      } else if (orderBy.field === 'org-verified') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).verified`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).verified`
-      } else if (orderBy.field === 'org-summary-mail-pass') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.pass`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.mail.pass`
-      } else if (orderBy.field === 'org-summary-mail-fail') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.fail`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.mail.fail`
-      } else if (orderBy.field === 'org-summary-mail-total') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.total`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.mail.total`
-      } else if (orderBy.field === 'org-summary-web-pass') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.pass`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.web.pass`
-      } else if (orderBy.field === 'org-summary-web-fail') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.fail`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.web.fail`
-      } else if (orderBy.field === 'org-summary-web-total') {
-        affiliationField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.total`
-        documentField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key).summaries.web.total`
-      } else if (orderBy.field === 'org-domain-count') {
-        affiliationField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key)._id claims RETURN e._to)`
-        documentField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._from).key)._id claims RETURN e._to)`
+        affiliationField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).userName`
+        documentField = aql`DOCUMENT(users, PARSE_IDENTIFIER(DOCUMENT(affiliations, ${beforeId})._to).key).userName`
       }
 
       beforeTemplate = aql`
         FILTER ${affiliationField} ${beforeTemplateDirection} ${documentField}
         OR (${affiliationField} == ${documentField}
-        AND FILTER TO_NUMBER(affiliation._key) > TO_NUMBER(${beforeId}))
+        AND TO_NUMBER(affiliation._key) < TO_NUMBER(${beforeId}))
       `
     }
   }
@@ -165,7 +67,7 @@ export const affiliationLoaderByOrgId = (
   let limitTemplate = aql``
   if (typeof first === 'undefined' && typeof last === 'undefined') {
     console.warn(
-      `User: ${userKey} did not have either \`first\` or \`last\` arguments set for: affiliationLoaderByOrgId.`,
+      `User: ${userKey} did not have either \`first\` or \`last\` arguments set for: affiliationConnectionLoaderByOrgId.`,
     )
     throw new Error(
       i18n._(
@@ -174,7 +76,7 @@ export const affiliationLoaderByOrgId = (
     )
   } else if (typeof first !== 'undefined' && typeof last !== 'undefined') {
     console.warn(
-      `User: ${userKey} attempted to have \`first\` and \`last\` arguments set for: affiliationLoaderByOrgId.`,
+      `User: ${userKey} attempted to have \`first\` and \`last\` arguments set for: affiliationConnectionLoaderByOrgId.`,
     )
     throw new Error(
       i18n._(
@@ -186,7 +88,7 @@ export const affiliationLoaderByOrgId = (
     if (first < 0 || last < 0) {
       const argSet = typeof first !== 'undefined' ? 'first' : 'last'
       console.warn(
-        `User: ${userKey} attempted to have \`${argSet}\` set below zero for: affiliationLoaderByOrgId.`,
+        `User: ${userKey} attempted to have \`${argSet}\` set below zero for: affiliationConnectionLoaderByOrgId.`,
       )
       throw new Error(
         i18n._(
@@ -197,7 +99,7 @@ export const affiliationLoaderByOrgId = (
       const argSet = typeof first !== 'undefined' ? 'first' : 'last'
       const amount = typeof first !== 'undefined' ? first : last
       console.warn(
-        `User: ${userKey} attempted to have \`${argSet}\` set to ${amount} for: affiliationLoaderByOrgId.`,
+        `User: ${userKey} attempted to have \`${argSet}\` set to ${amount} for: affiliationConnectionLoaderByOrgId.`,
       )
       throw new Error(
         i18n._(
@@ -213,15 +115,15 @@ export const affiliationLoaderByOrgId = (
     const argSet = typeof first !== 'undefined' ? 'first' : 'last'
     const typeSet = typeof first !== 'undefined' ? typeof first : typeof last
     console.warn(
-      `User: ${userKey} attempted to have \`${argSet}\` set as a ${typeSet} for: affiliationLoaderByOrgId.`,
+      `User: ${userKey} attempted to have \`${argSet}\` set as a ${typeSet} for: affiliationConnectionLoaderByOrgId.`,
     )
     throw new Error(
       i18n._(t`\`${argSet}\` must be of type \`number\` not \`${typeSet}\`.`),
     )
   }
 
-  let hasNextPageFilter = aql`FILTER TO_NUMBER(aff._key) > TO_NUMBER(LAST(retrievedAffiliations)._key)`
-  let hasPreviousPageFilter = aql`FILTER TO_NUMBER(aff._key) < TO_NUMBER(FIRST(retrievedAffiliations)._key)`
+  let hasNextPageFilter = aql`FILTER TO_NUMBER(affiliation._key) > TO_NUMBER(LAST(retrievedAffiliations)._key)`
+  let hasPreviousPageFilter = aql`FILTER TO_NUMBER(affiliation._key) < TO_NUMBER(FIRST(retrievedAffiliations)._key)`
   if (typeof orderBy !== 'undefined') {
     let hasNextPageDirection = aql``
     let hasPreviousPageDirection = aql``
@@ -236,85 +138,21 @@ export const affiliationLoaderByOrgId = (
     let affField, hasNextPageDocument, hasPreviousPageDocument
     /* istanbul ignore else */
     if (orderBy.field === 'user-username') {
-      affField = aql`DOCUMENT(users, PARSE_IDENTIFIER(aff._to).key).username`
-      hasNextPageDocument = aql`DOCUMENT(users, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._to).key).username`
-      hasPreviousPageDocument = aql`DOCUMENT(users, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._to).key).username`
-    } else if (orderBy.field === 'org-acronym') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).acronym`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).acronym`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).acronym`
-    } else if (orderBy.field === 'org-name') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).name`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).name`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).name`
-    } else if (orderBy.field === 'org-slug') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).slug`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).slug`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).slug`
-    } else if (orderBy.field === 'org-zone') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).zone`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).zone`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).zone`
-    } else if (orderBy.field === 'org-sector') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).sector`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).sector`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).sector`
-    } else if (orderBy.field === 'org-country') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).country`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).country`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).country`
-    } else if (orderBy.field === 'org-province') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).province`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).province`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).province`
-    } else if (orderBy.field === 'org-city') {
-      affField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).orgDetails).city`
-      hasNextPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).orgDetails).city`
-      hasPreviousPageDocument = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).orgDetails).city`
-    } else if (orderBy.field === 'org-verified') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).verified`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).verified`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).verified`
-    } else if (orderBy.field === 'org-summary-mail-pass') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.mail.pass`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.mail.pass`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.mail.pass`
-    } else if (orderBy.field === 'org-summary-mail-fail') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.mail.fail`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.mail.fail`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.mail.fail`
-    } else if (orderBy.field === 'org-summary-mail-total') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.mail.total`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.mail.total`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.mail.total`
-    } else if (orderBy.field === 'org-summary-web-pass') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.web.pass`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.web.pass`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.web.pass`
-    } else if (orderBy.field === 'org-summary-web-fail') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.web.fail`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.web.fail`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.web.fail`
-    } else if (orderBy.field === 'org-summary-web-total') {
-      affField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key).summaries.web.total`
-      hasNextPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key).summaries.web.total`
-      hasPreviousPageDocument = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key).summaries.web.total`
-    } else if (orderBy.field === 'org-domain-count') {
-      affField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(aff._from).key)._id claims RETURN e._to)`
-      hasNextPageDocument = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._from).key)._id claims RETURN e._to)`
-      hasPreviousPageDocument = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._from).key)._id claims RETURN e._to)`
+      affField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).userName`
+      hasNextPageDocument = aql`DOCUMENT(users, PARSE_IDENTIFIER(LAST(retrievedAffiliations)._to).key).userName`
+      hasPreviousPageDocument = aql`DOCUMENT(users, PARSE_IDENTIFIER(FIRST(retrievedAffiliations)._to).key).userName`
     }
 
     hasNextPageFilter = aql`
       FILTER ${affField} ${hasNextPageDirection} ${hasNextPageDocument}
       OR (${affField} == ${hasNextPageDocument}
-      AND TO_NUMBER(aff._key) > TO_NUMBER(LAST(retrievedAffiliations)._key))
+      AND TO_NUMBER(affiliation._key) > TO_NUMBER(LAST(retrievedAffiliations)._key))
     `
 
     hasPreviousPageFilter = aql`
       FILTER ${affField} ${hasPreviousPageDirection} ${hasPreviousPageDocument}
       OR (${affField} == ${hasPreviousPageDocument}
-      AND TO_NUMBER(aff._key) < TO_NUMBER(FIRST(retrievedAffiliations)._key))
+      AND TO_NUMBER(affiliation._key) < TO_NUMBER(FIRST(retrievedAffiliations)._key))
     `
   }
 
@@ -322,39 +160,7 @@ export const affiliationLoaderByOrgId = (
   if (typeof orderBy !== 'undefined') {
     /* istanbul ignore else */
     if (orderBy.field === 'user-username') {
-      sortByField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).username ${orderBy.direction},`
-    } else if (orderBy.field === 'org-acronym') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).acronym ${orderBy.direction},`
-    } else if (orderBy.field === 'org-name') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).name ${orderBy.direction},`
-    } else if (orderBy.field === 'org-slug') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).slug ${orderBy.direction},`
-    } else if (orderBy.field === 'org-zone') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).zone ${orderBy.direction},`
-    } else if (orderBy.field === 'org-sector') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).sector ${orderBy.direction},`
-    } else if (orderBy.field === 'org-country') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).country ${orderBy.direction},`
-    } else if (orderBy.field === 'org-province') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).province ${orderBy.direction},`
-    } else if (orderBy.field === 'org-city') {
-      sortByField = aql`TRANSLATE(${language}, DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).orgDetails).city ${orderBy.direction},`
-    } else if (orderBy.field === 'org-verified') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).verified ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-mail-pass') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.pass ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-mail-fail') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.fail ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-mail-total') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.mail.total ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-web-pass') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.pass ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-web-fail') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.fail ${orderBy.direction},`
-    } else if (orderBy.field === 'org-summary-web-total') {
-      sortByField = aql`DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key).summaries.web.total ${orderBy.direction},`
-    } else if (orderBy.field === 'org-domain-count') {
-      sortByField = aql`COUNT(FOR v, e IN 1..1 ANY DOCUMENT(organizations, PARSE_IDENTIFIER(affiliation._from).key)._id claims RETURN e._to) ${orderBy.direction},`
+      sortByField = aql`DOCUMENT(users, PARSE_IDENTIFIER(affiliation._to).key).userName ${orderBy.direction},`
     }
   }
 
@@ -384,19 +190,19 @@ export const affiliationLoaderByOrgId = (
     )
 
     LET hasNextPage = (LENGTH(
-      FOR aff IN affiliations
-        FILTER aff._key IN affiliationKeys
+      FOR affiliation IN affiliations
+        FILTER affiliation._key IN affiliationKeys
         ${hasNextPageFilter}
-        SORT ${sortByField} aff._key ${sortString} LIMIT 1
-        RETURN aff
+        SORT ${sortByField} affiliation._key ${sortString} LIMIT 1
+        RETURN affiliation
     ) > 0 ? true : false)
 
     LET hasPreviousPage = (LENGTH(
-      FOR aff IN affiliations
-        FILTER aff._key IN affiliationKeys
+      FOR affiliation IN affiliations
+        FILTER affiliation._key IN affiliationKeys
         ${hasPreviousPageFilter}
-        SORT ${sortByField} aff._key ${sortString} LIMIT 1
-        RETURN aff
+        SORT ${sortByField} affiliation._key ${sortString} LIMIT 1
+        RETURN affiliation
     ) > 0 ? true : false)
 
     RETURN {
@@ -410,7 +216,7 @@ export const affiliationLoaderByOrgId = (
     `
   } catch (err) {
     console.error(
-      `Database error occurred while user: ${userKey} was trying to query affiliations in affiliationLoaderByOrgId, error: ${err}`,
+      `Database error occurred while user: ${userKey} was trying to query affiliations in affiliationConnectionLoaderByOrgId, error: ${err}`,
     )
     throw new Error(i18n._(t`Unable to query affiliations. Please try again.`))
   }
@@ -420,7 +226,7 @@ export const affiliationLoaderByOrgId = (
     filteredAffiliations = await filteredAffiliationCursor.next()
   } catch (err) {
     console.error(
-      `Cursor error occurred while user: ${userKey} was trying to gather affiliations in affiliationLoaderByOrgId, error: ${err}`,
+      `Cursor error occurred while user: ${userKey} was trying to gather affiliations in affiliationConnectionLoaderByOrgId, error: ${err}`,
     )
     throw new Error(i18n._(t`Unable to load affiliations. Please try again.`))
   }
